This pull request includes various code refactoring and feature enhancements to improve the functionality and performance of the application. It introduces a global config module, updates the termination process of client processes, adds the ability to export failed accounts separately, improves logging and progress handling, optimizes client launching and communication, optimizes data loading, and introduces documentation and type annotations.

+++ client/connection/Connection.py
@@ -2,17 +2,26 @@ from requests import Session
 from requests import adapters
 import urllib3
 import subprocess
-from time import sleep
-from client.connection.credentials import getFreePort
 from client.connection.credentials import getAuthToken
+from client.connection.exceptions import LaunchFailedException
+from typing import List
 
-# inherited class for riot and league clients
 class Connection(Session):
-    def __init__(self):
-        # use our own port and auth token so we can connect to api while running headless
-        self._port = str(getFreePort())
+    """
+    Custom class for Riot and League clients' connection.
+
+    This class extends the `Session` class from the `requests` library to handle API requests to the client.
+    """
+    def __init__(self, port: int) -> None:
+        """
+        Initializes the Connection class with the specified port.
+
+        :param port: The port number used for the client connection.
+        """
+        self._port = str(port)
         self._url = "https://127.0.0.1:" + self._port
         self._authToken = getAuthToken()
+        self._process = None
         Session.__init__(self)
         retry = urllib3.util.retry.Retry(
             total = 5,
@@ -25,27 +34,41 @@ class Connection(Session):
         self.mount("https://", adapter)
         urllib3.disable_warnings()
 
-    # handles api requests to the client
-    def request(self, method, url, *args, **kwargs):
+    def request(self, method: str, url: str, *args, **kwargs) -> Session.request:
+        """
+        Sends an API request to the client.
+
+        :return: The API response.
+        """
         url = self._url + url
-        # setup riot authentication
+        # Setup Riot authentication
         kwargs["auth"] = "riot", self._authToken
         kwargs["verify"] = False
         
-        # return api request
+        # Return API request
         return Session.request(self, method, url, *args, **kwargs)
     
-    # opens a process (league/riot client) with given arguments
-    def getClient(self, processArgs):
+    def getClient(self, processArgs: List[str]) -> None:
+        """
+        Opens a process for the client application with the given arguments.
+
+        :param processArgs: The arguments for the client application process.
+
+        Raises:
+            LaunchFailedException: If launching the client application process fails.
+        """
         while True:
             try:
                 self._process = subprocess.Popen(processArgs)
                 return
-            except:
-                sleep(1)
+            except OSError as error:
+                raise LaunchFailedException(self.__class__.__name__, error)
     
-    # terminates client process 
-    def __del__(self):
-        self._process.terminate()
+    def __del__(self) -> None:
+        """
+        Terminates the client process.
+        """
+        if self._process is not None:
+            self._process.terminate()
